fel: add "reset" command to trigger watchdog reset
On some occasions it would be nice to be able to reset a board via a
FEL command. One use case is writing the SPI flash, then automatically
booting into the newly installed firmware. Also it may be useful for
boards without a reset button.
Add a "reset" command to sunxi-fel, which uploads a few instructions.
They program the watchdog for system reset within the shortest possible
period, then just wait in a WFI loop.
So far there are two types of watchdog IP, one "older" used on the A10,
A13, A20 and R40, and a "newer" version used on all the other SoCs.
Add the watchdog MMIO base address and a flag to denote the type to
our soc_info table, and let the reset command use this information
accordingly.
Tested on A64, H5, H2+ and A20.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>