fel: prepare SPL write for raw transport
At the moment our SPL upload routine requires a valid bootable SPL/boot0
image, with an eGON header. This is indeed what most people expect, and
the rigorous header and checksum checks both in sunxi-fel and on the
device help to avoid confusion due to booting the wrong image.

However an eGON header is technically speaking not strictly necessary,
and just running arbitrary code in SRAM opens two use cases:
- People experimenting with code can just upload a normal binary,
  without the need to add an eGON header.
- It allows to upload manipulated eGON binaries, for instance to trick
  SPLs into booting from other boot sources than FEL.

Extend aw_fel_write_and_execute_spl() to take a parameter which triggers
those checks. Setting "spl" to false will skip the eGON header checks,
and uploads a simplified version of the fel-to-spl-thunk, which skips
the on-device header checks and avoids the signature manipulation.
We still observe SPL image size limits, and do all the required buffer
swapping.

Signed-off-by: Andre Przywara <osp@andrep.de>