fel: add "spiflash-boot" command
Add a command to boot a device from its SPI flash. This will read the
SPL from the beginning of the SPI flash, fake the boot source indicator
to 0x03, upload it to SRAM and then execute it.
This mimics what the BootROM does when booting from SPI flash, and
U-Boot's SPL will happily continue loading further data from SPI flash
and execute it.

This allows to boot directly into a device as if it has been booted from
SPI flash, even when higher priority boot media would otherwise prevent
that. Also it helps to install firmware to SPI flash and boot it right
away.

Signed-off-by: Andre Przywara <osp@andrep.de>