fel: add gpio command
Add a sunxi-fel command to either query the input level of a GPIO pin,
or set that output level.
For example "sunxi-fel gpio input PB24" would print either 0 or 1,
depending on whether the input level on pin PB24 is low or high. The
respective pin is configured as a GPIO input first.
"sunxi-fel gpio toggle PB24" would program that pin as a GPIO output and
change its level, "set" would force it HIGH, "clear" would force it LOW.

At the moment this only supports the primary GPIO controller (ports A-J).

Signed-off-by: Andre Przywara <osp@andrep.de>